+++ src/app/home/home.component.ts
@@ -4,7 +4,8 @@ import { UserService } from '../shared/user.service';
 import { Injectable} from '@angular/core';
 import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
-import { RootObject, Coin } from '../classes/Coins';
+import { RootObject } from "../models/RootObject";
+import { Coin } from "../models/Coin";
 import { coinService } from '../shared/coin.service';
 import * as CanvasJS from '../JS/canvasjs.min';
 import * as $ from "jquery";
