+++ src/app/models/CoinsNews.ts
@@ -1,0 +1,6 @@
+import { Article } from "./Article";
+export class CoinsNews {
+    status: string;
+    totalResults: number;
+    articles: Article[];
+}
