+++ src/app/models/Platform.ts
@@ -1,0 +1,7 @@
+export class Platform {
+    id: number;
+    name: string;
+    symbol: string;
+    slug: string;
+    token_address: string;
+}
