+++ src/app/models/Quote.ts
@@ -1,0 +1,4 @@
+import { Usd } from './Usd';
+export class Quote {
+    usd: Usd = new Usd();
+}
