+++ src/app/models/RootObject.ts
@@ -1,0 +1,6 @@
+import { Status } from './Status';
+import { Coin } from './Coin';
+export class RootObject {
+    status: Status;
+    data: Coin[];
+}
