+++ src/app/models/Source.ts
@@ -1,0 +1,4 @@
+export class Source {
+    id: string;
+    name: string;
+}
