+++ src/app/models/Status.ts
@@ -1,0 +1,8 @@
+export class Status {
+    timestamp: Date;
+    error_code: number;
+    error_message?: any;
+    elapsed: number;
+    credit_count: number;
+    notice?: any;
+}
