+++ src/app/models/Usd.ts
@@ -1,0 +1,9 @@
+export class Usd {
+    price: number;
+    volume_24h: number;
+    percent_change_1h: number;
+    percent_change_24h: number;
+    percent_change_7d: number;
+    market_cap: number;
+    last_updated: Date;
+}
