+++ src/app/shared/coin.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from "@angular/core"
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { RootObject, Article } from '../classes/Coins';
+import { Article } from "../models/Article";
+import { RootObject } from "../models/RootObject";
 
  @Injectable({
      providedIn: 'root'
